Api/requirements.txt
altgraph==0.17
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
ecdsa==0.14.1
future==0.18.2
idna==2.10
joblib==0.17.0
Markdown==3.3.4
MouseInfo==0.1.3
nltk==3.5
oauthlib==3.1.0
pefile==2019.4.18
Pillow==8.1.0
playsound==1.2.2
psycopg2==2.8.6
pyasn1==0.4.8
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
pyinstaller-hooks-contrib==2020.11
PyJWT==1.7.1
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-jose==3.2.0
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
secure-smtplib==0.1.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
tqdm==4.54.1
urllib3==1.26.4
